Add placeholder component to domains/add flow
